Initial FFI tests
I would like to test if it is possible to ship a embedded Dat engine
into a mobile app.

Talking to friends, I've learned that iOS does not ship with any
`exec`-like API, so it would not be viable to ship a binary there.

After taking a look on Dart FFI, I'm considering testing it directly
with FFI.
The effort to build a Flutter + binary application seems to be on the
same league as pushing a Flutter + ffi sample.

This test intends to do the following:

- [ ] Exec a hardcoded sync from Flutter Android
- [ ] Validate isolate FFI in the background
- [ ] Test it on desktop, if possible.
- [ ] Test on the iOS simulator (and maybe an iPhone if I can find a volunteer)

The steps of evolving this sample would be, IMO:

- Sync a hardcoded dat
- Sync a dat url passed over FFI, each on an isolate
- Port [hyperdrive-daemon](https://github.com/andrewosh/hyperdrive-daemon) and embed a server into the app, to avoid FFI method calls