Converts the locator from a stream of bytes, to an stream of SocketAddrs
SocketAddrs contains a pair of both an Ip (and version) and the port
reported.

We can now locate a peer on our local network, as we can already filter
the wrong packets from the mdns broadcasts.

This stream will return a sequence of Hyperswarm peers with their
reported port, so we can locate and connect.