Support handling multiple feeds on MDNS discovery
This commit allow us to add and remove feed itens from the mDNS
discovery stack.

We are not capable of integrating with @Frando's `hypercore-replicator`,
which should let us introduce new discovery feeds when looking on mdns.

The binary example tests that we can still have a stream of items, and
that we can add and remove topics in the background.

Ideally, we could later move into the "subscribers" direction, instead
of a single stream, similar to how @Frando is doing on the `hyperspace`
code. So we can avoid wrapping things on a `Arc<RwLock<_>>`.

TODO: Refactor Hyperstack - possibly using `hypercore-replicator` and
move out the `hyperdrive` into it's own crate.

Closes #22